+++ backend/controllers/SupplierController.php
@@ -63,16 +63,22 @@ class SupplierController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Supplier();
-		$kode_generate = $model->generateKodeSupplier();
-		$model->kode_supplier = $kode_generate;
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
-        }
+		if(Yii::$app->user->can('tambahSupplier')){
+			$model = new Supplier();
+			$kode_generate = $model->generateKodeSupplier();
+			$model->kode_supplier = $kode_generate;
+			if ($model->load(Yii::$app->request->post()) && $model->save()) {
+				return $this->redirect(['view', 'id' => $model->id]);
+			} else {
+				return $this->render('create', [
+					'model' => $model,
+				]);
+			}
+		}else{
+			Yii::$app->getSession()->setFlash('error', 'Anda tidak mempunyai hak untuk menambah data');
+			return $this->redirect(['index']);
+		}
+        
     }
 
     /**
@@ -83,15 +89,21 @@ class SupplierController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+		if(Yii::$app->user->can('updateSupplier')){
+			$model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
-        }
+			if ($model->load(Yii::$app->request->post()) && $model->save()) {
+				return $this->redirect(['view', 'id' => $model->id]);
+			} else {
+				return $this->render('update', [
+					'model' => $model,
+				]);
+			}
+		}else{
+			Yii::$app->getSession()->setFlash('error', 'Anda tidak mempunyai hak akses untuk melakukan perubahan data');
+			return $this->redirect(['index']);
+		}
+        
     }
 
     /**
@@ -102,9 +114,15 @@ class SupplierController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+		if(Yii::$app->user->can('deleteSupplier')){
+			$this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+			return $this->redirect(['index']);
+		}else{
+			Yii::$app->getSession()->setFlash('error', 'Anda tidak mempunyai hak untuk melakukan penghapusan data');
+			return $this->redirect(['index']);
+		}
+        
     }
 
     /**
